+++ src/main/java/Website/ProductInfo.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 public class ProductInfo {
     String name;
     String description;
-    Float price;
+    Double price;
     Integer quantity;
     String category;
     Boolean limited;
